counting sort

+++ sort.h
@@ -27,7 +27,7 @@ void selection_sort(int *array, int size);
 void quick_sort(int *array, int size);
 void shell_sort(int *array, size_t size);
 void cocktail_sort_list(listint_t **list);
-void counting_sort(int *array, int size);
+void counting_sort(int *array, size_t size);
 void merge_sort(int *array, int size);
 void heap_sort(int *array, int size);
 void radix_sort(int *array, int size);
